Move errors back to interfaces.py
