+++ warehouse/packaging/models.py
@@ -1154,34 +1154,3 @@ class AlternateRepository(db.Model):
     name: Mapped[str]
     url: Mapped[str]
     description: Mapped[str]
-
-
-class ProjectNameUnavailableInvalid:
-    pass
-
-
-class ProjectNameUnavailableStdlib:
-    pass
-
-
-class ProjectNameUnavailableExisting:
-    def __init__(self, existing_project: Project):
-        self.existing_project: Project = existing_project
-
-
-class ProjectNameUnavailableProhibited:
-    pass
-
-
-class ProjectNameUnavailableSimilar:
-    def __init__(self, similar_project: Project):
-        self.similar_project: Project = similar_project
-
-
-ProjectNameUnavailableError = (
-    ProjectNameUnavailableInvalid
-    | ProjectNameUnavailableStdlib
-    | ProjectNameUnavailableExisting
-    | ProjectNameUnavailableProhibited
-    | ProjectNameUnavailableSimilar
-)
